+++ apps/studio/components/interfaces/Billing/Payment/PaymentConfirmation.tsx
@@ -1,0 +1,36 @@
+import { useStripe } from '@stripe/react-stripe-js'
+import { PaymentIntentResult } from '@stripe/stripe-js'
+import { useEffect } from 'react'
+import { LoadingLine } from 'ui'
+
+export const PaymentConfirmation = ({
+  paymentIntentSecret,
+  onPaymentIntentConfirm,
+  onLoadingChange,
+  paymentMethodId,
+}: {
+  paymentIntentSecret: string
+  paymentMethodId: string
+  onPaymentIntentConfirm: (response: PaymentIntentResult) => void
+  onLoadingChange: (loading: boolean) => void
+}) => {
+  const stripe = useStripe()
+
+  useEffect(() => {
+    if (stripe && paymentIntentSecret) {
+      onLoadingChange(true)
+      stripe!
+        .confirmCardPayment(paymentIntentSecret, { payment_method: paymentMethodId })
+        .then((res) => {
+          onPaymentIntentConfirm(res)
+          onLoadingChange(false)
+        })
+        .catch((err) => {
+          console.error(err)
+          onLoadingChange(false)
+        })
+    }
+  }, [paymentIntentSecret, stripe])
+
+  return <LoadingLine loading={true} />
+}
