+++ apps/studio/components/interfaces/Organization/BillingSettings/CreditTopUp.tsx
@@ -1,6 +1,6 @@
 import HCaptcha from '@hcaptcha/react-hcaptcha'
 import { zodResolver } from '@hookform/resolvers/zod'
-import { Elements, useStripe } from '@stripe/react-stripe-js'
+import { Elements } from '@stripe/react-stripe-js'
 import { loadStripe, PaymentIntentResult } from '@stripe/stripe-js'
 import { PermissionAction, SupportCategories } from '@supabase/shared-types/out/constants'
 import { useQueryClient } from '@tanstack/react-query'
@@ -35,10 +35,10 @@ import {
   Form_Shadcn_,
   FormField_Shadcn_,
   Input_Shadcn_,
-  LoadingLine,
 } from 'ui'
 import { FormItemLayout } from 'ui-patterns/form/FormItemLayout/FormItemLayout'
 import PaymentMethodSelection from './Subscription/PaymentMethodSelection'
+import { PaymentConfirmation } from 'components/interfaces/Billing/Payment/PaymentConfirmation'
 
 const stripePromise = loadStripe(STRIPE_PUBLIC_KEY)
 
@@ -330,35 +330,4 @@ export const CreditTopUp = ({ slug }: { slug: string | undefined }) => {
       </Dialog>
     </div>
   )
-}
-
-const PaymentConfirmation = ({
-  paymentIntentSecret,
-  onPaymentIntentConfirm,
-  onLoadingChange,
-  paymentMethodId,
-}: {
-  paymentIntentSecret: string
-  paymentMethodId: string
-  onPaymentIntentConfirm: (response: PaymentIntentResult) => void
-  onLoadingChange: (loading: boolean) => void
-}) => {
-  const stripe = useStripe()
-
-  useEffect(() => {
-    if (stripe && paymentIntentSecret) {
-      onLoadingChange(true)
-      stripe!
-        .confirmCardPayment(paymentIntentSecret, { payment_method: paymentMethodId })
-        .then((res) => {
-          onPaymentIntentConfirm(res)
-          onLoadingChange(false)
-        })
-        .catch((err) => {
-          onLoadingChange(false)
-        })
-    }
-  }, [paymentIntentSecret, stripe])
-
-  return <LoadingLine loading={true} />
 }
